矿机操作脚本：支持从OOLuaHelper::opt()获取矿机的登录名和密码。

+++ src/lua/scripts/minerRebooter/antminer-http-cgi.lua
@@ -48,15 +48,22 @@ function configurator.doMakeResult(context, response, stat)
     
     if (step == "auth") then
 		if (response.statCode == "401") then
-			local request = http.parseRequest(context:requestContent())
-			local requestContent, err = http.makeAuthRequest(request, response, 'root', 'root')
-			
-			if (err) then
-				context:setStepName("end")
-				miner:setStat('failed: ' .. err)
-			else
-				context:setStepName("reboot")
-				context:setRequestContent(requestContent)
+			loginPassword = utils.getMinerLoginPassword(miner:fullTypeStr())
+            
+            if (loginPassword == nil) then
+                context:setStepName("end")
+                context:miner():setStat("require password")
+            else
+				local request = http.parseRequest(context:requestContent())
+				local requestContent, err = http.makeAuthRequest(request, response, loginPassword.userName, loginPassword.password)
+				
+				if (err) then
+					context:setStepName("end")
+					miner:setStat('failed: ' .. err)
+				else
+					context:setStepName("reboot")
+					context:setRequestContent(requestContent)
+				end
 			end
 		else
 			context:setStepName("end")
